+++ main/java/unstudio/chinacraft/item/ItemSuperBow.java
@@ -41,10 +41,10 @@ public class ItemSuperBow extends ItemBow{
 
         if (flag || entityPlayer.inventory.hasItem(Items.arrow))
         {
-            float f = (float)j / 20.0F;
+            float f = j / 20.0F;
             f = (f * f + f * 2.0F) / 3.0F;
 
-            if ((double)f < 0.1D)
+            if (f < 0.1D)
             {
                 return;
             }
@@ -65,7 +65,7 @@ public class ItemSuperBow extends ItemBow{
 
             if (k > 0)
             {
-                entityarrow.setDamage(entityarrow.getDamage() + (double)k * 0.5D + 0.5D);
+                entityarrow.setDamage(entityarrow.getDamage() + k * 0.5D + 0.5D);
             }
 
             int l = EnchantmentHelper.getEnchantmentLevel(Enchantment.punch.effectId, itemStack);
